Finished api with .jar executable

+++ src/main/java/com/tienda/service/ProductService.java
@@ -4,14 +4,15 @@ import java.util.List;
 import java.util.function.Supplier;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.tienda.main.RepositoryInterface;
 import com.tienda.model.ProductModel;
-
-import jakarta.transaction.Transactional;
+ 
 @Service
 public class ProductService {
     @Autowired
@@ -28,33 +29,41 @@ public class ProductService {
        
     }
     @Transactional
-    public void updateProduct(int id,String name, String description, int amount, int price){
+    public void updateProduct(int id,String name, String description, Integer amount, Integer price){
         
        if(!rInterface.existsById(id)){
         throw new IllegalStateException("no existe producto con id "+id);
        }
        ProductModel tempProduct=rInterface.findById(id).get();
-       if (name!=null && name.length()>0) {tempProduct.setName(name);}
-       if (description!=null && name.length()>0) {tempProduct.setDescription(description);}
+       if (name.length()>0) {tempProduct.setName(name);}
+       if (description.length()>0) {tempProduct.setDescription(description);}
+     
+ 
        if (amount>-1) {tempProduct.setAmount(amount);}
        if (price>-1) {tempProduct.setPrice(price);}
        rInterface.save(tempProduct);
 
     }
     public void deleteProduct(int id) {
-    //    if (!rInterface.existsById(id)) {
-    //     throw new IllegalStateException("no existe producto con id "+id);
-    //    }
+       if (!rInterface.existsById(id)) {
+        throw new IllegalStateException("no existe producto con id "+id);
+       }
         rInterface.deleteById(id);
     }
 
     public List<ProductModel> getProductbyName(String name) {
         return List.of(rInterface.findByName(name));
-        // return List.of(rInterface.findById(name).orElseThrow(() -> new IllegalStateException("no existe producto con el id "+id)));
     }
 
     public List<ProductModel> getProductbyId(int id) {
         return List.of(rInterface.findById(id).orElseThrow(() -> new IllegalStateException("no existe producto con el id "+id)));
     }
+
+
+	public List<ProductModel> getSortedPrice() {
+
+
+		return rInterface.findAll(Sort.by(Sort.Direction.DESC, "price"));
+	}
     
 }
